+++ src/tools/getBuildFailureLogs/tool.ts
@@ -8,8 +8,11 @@ export const getBuildFailureLogsTool = {
     Input options (EXACTLY ONE of these two options must be used):
 
     Option 1 - Direct URL (provide ONE of these):
-    - failedJobURL: The URL of the failed CircleCI job (must be provided by the user)
-    - failedPipelineURL: The URL of the failed CircleCI pipeline (must be provided by the user)
+    - projectURL: The URL of the CircleCI project in any of these formats:
+      * Project URL: https://app.circleci.com/pipelines/gh/organization/project
+      * Pipeline URL: https://app.circleci.com/pipelines/gh/organization/project/123
+      * Workflow URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def
+      * Job URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def/jobs/xyz
 
     Option 2 - Project Detection (ALL of these must be provided together):
     - workspaceRoot: The absolute path to the workspace root