+++ src/tools/getFlakyTests/inputSchema.ts
@@ -19,7 +19,12 @@ export const getFlakyTestLogsInputSchema = z.object({
   projectURL: z
     .string()
     .describe(
-      'The URL of the CircleCI project. This should be a link to the project in the CircleCI web app.',
+      'The URL of the CircleCI project. Can be any of these formats:\n' +
+        '- Project URL: https://app.circleci.com/pipelines/gh/organization/project\n' +
+        '- Project URL with branch: https://app.circleci.com/pipelines/gh/organization/project?branch=feature-branch\n' +
+        '- Pipeline URL: https://app.circleci.com/pipelines/gh/organization/project/123\n' +
+        '- Workflow URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def\n' +
+        '- Job URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def/jobs/xyz',
     )
     .optional(),
 });