+++ src/tools/getFlakyTests/tool.ts
@@ -9,8 +9,12 @@ export const getFlakyTestLogsTool = {
 
     Input options (EXACTLY ONE of these two options must be used):
 
-    Option 1 - Direct URL:
-    - projectURL: The URL of the CircleCI project (must be provided by the user)
+    Option 1 - Direct URL (provide ONE of these):
+    - projectURL: The URL of the CircleCI project in any of these formats:
+      * Project URL: https://app.circleci.com/pipelines/gh/organization/project
+      * Pipeline URL: https://app.circleci.com/pipelines/gh/organization/project/123
+      * Workflow URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def
+      * Job URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def/jobs/xyz
 
     Option 2 - Project Detection (ALL of these must be provided together):
     - workspaceRoot: The absolute path to the workspace root